sm8150-common: adjustments for AUX support
Add modded GCam package support for AUX camera to the packagelist.
Additionally add the OnePlus Camera to priv-app list, if it gets installed (although might be not needed)
At last add the permissions for OnePlus Camera and Gallery, if they get installed as priv-apps (usually only needed when added as prebuilts though).