sm8150-common: Fix OTA auto-flashing with encrypted f2fs userdata
If userdata is on f2fs and encrypted (Stock OOS behaviour for userdata), the /cache/recovery/block.map is not created due to a denial.

Due to this, the automatic flashing of the OTA without user
interaction fails and the user needs to manually mount data, and flash the OTA manually.
This might be a rare case on our devices, but we should take care of it...